requirements.txt
Click>=6.0
click-log>=0.2.1
pyserial==3.4
simplejson>=3.6.0
pyzmq
schema
